requirements.txt
langgraph==0.2.55
langchain-core==0.3.28
groq==0.9.0
chromadb==0.5.23
sentence-transformers==2.7.0
torch==2.2.2
torchvision==0.17.2
streamlit==1.38.0
python-dotenv==1.0.1
google-genai>=0.3.0